+++ tests/integration/test_kernel_maker.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:cc1f8e944fffc129c5d989b127a0508f62dc25348848c2f16a34265f07e923ae
-size 14099
+oid sha256:fe08dd85764d4c794e345aa5ab10a66ce4226418f346971a7726b4bac47724fe
+size 9821