+++ tests/integration/test_tier0_geolocation.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:345026ac2c42a4313ab5872637a1016e0620bebd5e06466c98a5d1d627c8dfd8
-size 5846
+oid sha256:e6e6fc5a6f12e5c48f479f03c2ea0f490ce77982cd937bfdca92f303cfdcc47f
+size 5720