+++ tests/plugins/spice_fixtures.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:0e94cf6eb1eb1730265079f9d7862909d0d6c7350f6b30e95f61e145ccb789f7
-size 3630
+oid sha256:b2b70b2286e8c8a8650e67f2fc45e63afb24903c4bda845b374432b92b4aeae0
+size 3701