+++ libera_utils/aws/processing_step_function_trigger.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f922d98a38ebac5eca3f6850df371e62c2b8f46a5683eabb3d173ec7264e6f3a
-size 5170
+oid sha256:484dc2f3e255c4daea32f1d5ce0a58a1ddec1ffc6831b74fa4f39e5be336df7b
+size 5166