+++ libera_utils/io/filenaming.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:e08cdc9bef17a4b7f2773033e7fb3c70c21bc92214d9a59df4e22f064a83e73d
-size 35662
+oid sha256:cece320b46da864ad54aef28d82cfd32e0b75376f96348f6b4962b27ba0e2f5a
+size 25419