+++ libera_utils/io/manifest.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:802bb446708bbaded510c06c6f83d5dafee43d8b00f5213a6bec43f484193786
-size 11490
+oid sha256:daea5f6b3c4859ede06474b2e3ab6a51c54a624ece6fa9a7a1fc4661445ff4eb
+size 11486