+++ tests/integration/test_kernel_maker.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f28f5c577bac5d94e8a563f6b38730b9eb7c0a6b062c6f83666e1024d0c6bc35
-size 11426
+oid sha256:991c31427342ae0ff0da8968644a4c677af184f3dcf5a0bc7ae6f5939c00f120
+size 11614