+++ tests/plugins/manifest_fixtures.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:fe96dc3081f32daa69805414554f65b394b914df48c970d609bb904f0bc04a98
-size 1698
+oid sha256:19b3c7e00c879f1a817f0b7db70242c9fad48cb0e98b4c93b06841101f9974a4
+size 1694