+++ tests/unit/test_aws/test_constants.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:8000f166658608e45b42a4f9b59565f1846259167e3877e1b89fac6e011ce19f
-size 3596
+oid sha256:7e74fa3dd6b8c68e8af8751b41501410b88d1a07ee5c8ea370820afe02025e54
+size 31