+++ tests/unit/test_aws/test_ecr_upload.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:72b3767d3af14e92a1706c0c5f7975471ab9d39b4c33174e74148e186b340f2d
-size 11686
+oid sha256:b6da28f13964d868f769c6ba22a41e0a7af3874592a13571bee3e939f01b3fbd
+size 11682