+++ tests/unit/test_aws/test_processing_step_function_trigger.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:06f009b78e46afd110c941eaa6ec76002941d99a19bf71eb1ca43a0417da8bd4
-size 2779
+oid sha256:c5b8a3b74ef76c3c3c8a8b915c7c9cf807fe38d3acd99592ceeaca40cd598b6f
+size 2775