+++ tests/unit/test_io/test_manifest.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:077f3f3e2ad3d7bb3fadb6f29d731913fede3d5be08fa3ad198f465fdeb10ebd
-size 9738
+oid sha256:30627268910392200830ace5d7f9e815e67740d030d588550b9376219c429311
+size 9734