+++ doc/source/changelog.md
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:cc070984d2e0421989d69bbf9a27403db07fe6081654f85abe31116ac9ca3753
-size 8149
+oid sha256:2cbcb9d94d2cee6647b64ceec945a652ccb4c677e5072a6739c49804ee9a2892
+size 8321