+++ doc/source/changelog.md
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:2cbcb9d94d2cee6647b64ceec945a652ccb4c677e5072a6739c49804ee9a2892
-size 8321
+oid sha256:3e70d5fac78ef5845065649ae5879c68217ddb4faf5049ff49a48df56cee8b2e
+size 9349